+++ tests/integration/__init__.py
@@ -2,5 +2,5 @@
 Package for integration tests
 """
 
-#TODO Use built-in BlueSky scenarios for testing initially. Once we can upload custom scenarios
+# TODO Use built-in BlueSky scenarios for testing initially. Once we can upload custom scenarios
 # then we can define custom ones for each integration test.